requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# Memory and vector storage
mem0ai>=0.1.8
qdrant-client>=1.9.0

# ModelScope integration
modelscope>=1.17.0
transformers>=4.40.0
torch>=2.2.0

# HTTP and API
requests>=2.31.0
aiohttp>=3.9.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# LangSmith for tracing and debugging
langsmith>=0.1.0

# Logging and utilities
rich>=13.7.0
colorama>=0.4.6

# Web interface
streamlit>=1.28.0
streamlit-chat>=0.1.0